+++ nyaacli/torrenting.py
@@ -49,7 +49,7 @@ def download_torrent(filename: str, result_filename: str = None, show_progress: 
             alert: libtorrent.alert
             for alert in alerts:
                 if alert.category() & libtorrent.alert.category_t.error_notification:
-                    sys.stdout.write(f"{red('[Alert]')} {alert} \r")
+                    logger.debug(f"{red('[Alert]')} {alert}")
 
             sys.stdout.flush()