+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Removed off-switch for `--debug` flag (used to be `--no-debug`). It's already off by default so it was an useless argument
 
+- Add option to sort by different values other than seeders with `--sort-by` (`-s`). (Closes Issue [`#27`](https://github.com/johnvictorfs/nyaa-cli/issues/27)). See linked issue to see options available.
+
 ---
 
 ## 0.1.8