+++ CHANGELOG.md
@@ -2,9 +2,23 @@
 
 ---
 
+## 0.2.0
+
+- Only show torrent alert messages in debug mode (`--debug`) (Closes Issue [`#25`](https://github.com/johnvictorfs/nyaa-cli/issues/25))
+
+- Add argument to search for only trusted torrents `-t` (`--trusted`) (Closes Issue [`#26`](https://github.com/johnvictorfs/nyaa-cli/issues/26))
+
+- Add argument alias for `--debug` (`-d`)
+
+- Removed off-switch for `--debug` flag (used to be `--no-debug`). It's already off by default so it was an useless argument
+
+- Add option to sort by different values other than seeders with `--sort-by` (`-s`). (Closes Issue [`#27`](https://github.com/johnvictorfs/nyaa-cli/issues/27)). See linked issue to see options available.
+
+---
+
 ## 0.1.8
 
-- Add argument `-n` (`--number`, default = 10) to specify the number of entries to select from 
+- Add argument `-n` (`--number`, default = 10) to specify the number of entries to select from ([`#24`](https://github.com/johnvictorfs/nyaa-cli/commit/53771685f94f2d34c257b45c2ca749b08ab18ac2))
 
 ---